refactor(backend): retarget cutover tests to ports and split services
