+++ apps/backend/src/infrastructure/external/spotify-playlist-library.service.test.ts
@@ -136,5 +136,204 @@ describe("SpotifyPlaylistLibraryService.getMyPlaylists", () => {
     const playlists = await buildService().getMyPlaylists();
 
     expect(playlists.map((playlist) => playlist.id)).toEqual(["owned", "accessible"]);
+    expect(fetchMock).not.toHaveBeenCalledWith(
+      expect.stringContaining("/playlists/owned/items"),
+      expect.anything(),
+    );
+    expect(fetchMock).toHaveBeenCalledWith(
+      expect.stringContaining("/playlists/accessible/items"),
+      expect.anything(),
+    );
+    expect(fetchMock).toHaveBeenCalledWith(
+      expect.stringContaining("/playlists/inaccessible/items"),
+      expect.anything(),
+    );
+  });
+
+  it("uses the access cache to avoid repeated non-owned playlist probes", async () => {
+    const accessiblePlaylist = buildPlaylist({
+      id: "accessible",
+      name: "Accessible",
+      external_urls: { spotify: "https://open.spotify.com/playlist/accessible" },
+    });
+
+    const fetchMock = vi.fn(async (input: string | URL) => {
+      const url = input.toString();
+
+      if (url === "https://api.spotify.com/v1/me") {
+        return jsonResponse({ id: "current-user" });
+      }
+
+      if (url === "https://api.spotify.com/v1/me/playlists?limit=50") {
+        return jsonResponse({ items: [accessiblePlaylist], next: null });
+      }
+
+      if (url.includes("/playlists/accessible/items")) {
+        return jsonResponse({ total: 1 });
+      }
+
+      throw new Error(`Unexpected Spotify URL: ${url}`);
+    });
+    vi.stubGlobal("fetch", fetchMock);
+
+    const service = buildService();
+    await service.getMyPlaylists();
+    await service.getMyPlaylists();
+
+    const itemProbeCalls = fetchMock.mock.calls.filter(([input]) =>
+      input.toString().includes("/playlists/accessible/items"),
+    );
+    expect(itemProbeCalls).toHaveLength(1);
+  });
+
+  it("uses probe total when a non-owned playlist item is missing track total", async () => {
+    const accessiblePlaylist = buildPlaylist({
+      id: "accessible",
+      name: "Accessible",
+      tracks: {},
+      external_urls: { spotify: "https://open.spotify.com/playlist/accessible" },
+    });
+
+    const fetchMock = vi.fn(async (input: string | URL) => {
+      const url = input.toString();
+
+      if (url === "https://api.spotify.com/v1/me") {
+        return jsonResponse({ id: "current-user" });
+      }
+
+      if (url === "https://api.spotify.com/v1/me/playlists?limit=50") {
+        return jsonResponse({ items: [accessiblePlaylist], next: null });
+      }
+
+      if (url.includes("/playlists/accessible/items")) {
+        return jsonResponse({ total: 42 });
+      }
+
+      throw new Error(`Unexpected Spotify URL: ${url}`);
+    });
+    vi.stubGlobal("fetch", fetchMock);
+
+    const playlists = await buildService().getMyPlaylists();
+
+    expect(playlists).toHaveLength(1);
+    expect(playlists[0]?.tracks).toBe(42);
+  });
+
+  it("expires cached access probe results and re-probes deterministically", async () => {
+    const accessiblePlaylist = buildPlaylist({
+      id: "accessible",
+      name: "Accessible",
+      external_urls: { spotify: "https://open.spotify.com/playlist/accessible" },
+    });
+
+    const fetchMock = vi.fn(async (input: string | URL) => {
+      const url = input.toString();
+
+      if (url === "https://api.spotify.com/v1/me") {
+        return jsonResponse({ id: "current-user" });
+      }
+
+      if (url === "https://api.spotify.com/v1/me/playlists?limit=50") {
+        return jsonResponse({ items: [accessiblePlaylist], next: null });
+      }
+
+      if (url.includes("/playlists/accessible/items")) {
+        return jsonResponse({ total: 1 });
+      }
+
+      throw new Error(`Unexpected Spotify URL: ${url}`);
+    });
+    vi.stubGlobal("fetch", fetchMock);
+
+    const service = buildService();
+    await service.getMyPlaylists();
+
+    const cache = (
+      service as unknown as {
+        playlistAccessCache: Map<string, { expiresAt: number }>;
+      }
+    ).playlistAccessCache;
+    const entry = cache.get("accessible");
+    expect(entry).toBeDefined();
+    entry!.expiresAt = Date.now() - 1;
+
+    await service.getMyPlaylists();
+
+    const itemProbeCalls = fetchMock.mock.calls.filter(([input]) =>
+      input.toString().includes("/playlists/accessible/items"),
+    );
+    expect(itemProbeCalls).toHaveLength(2);
+  });
+
+  it("degrades to owned and cached accessible playlists during Spotify probe 429 cooldown", async () => {
+    const ownedPlaylist = buildPlaylist({
+      id: "owned",
+      name: "Owned",
+      owner: { ...buildPlaylist().owner, id: "current-user", display_name: "Me" },
+      external_urls: { spotify: "https://open.spotify.com/playlist/owned" },
+    });
+    const cachedAccessiblePlaylist = buildPlaylist({
+      id: "cached-accessible",
+      name: "Cached Accessible",
+      external_urls: { spotify: "https://open.spotify.com/playlist/cached-accessible" },
+    });
+    const rateLimitedPlaylist = buildPlaylist({
+      id: "rate-limited",
+      name: "Rate Limited",
+      external_urls: { spotify: "https://open.spotify.com/playlist/rate-limited" },
+    });
+    const unknownPlaylist = buildPlaylist({
+      id: "unknown",
+      name: "Unknown",
+      external_urls: { spotify: "https://open.spotify.com/playlist/unknown" },
+    });
+
+    const fetchMock = vi.fn(async (input: string | URL) => {
+      const url = input.toString();
+
+      if (url === "https://api.spotify.com/v1/me") {
+        return jsonResponse({ id: "current-user" });
+      }
+
+      if (url === "https://api.spotify.com/v1/me/playlists?limit=50") {
+        return jsonResponse({
+          items: [ownedPlaylist, cachedAccessiblePlaylist, rateLimitedPlaylist, unknownPlaylist],
+          next: null,
+        });
+      }
+
+      if (url.includes("/playlists/cached-accessible/items")) {
+        return jsonResponse({ total: 1 });
+      }
+
+      if (url.includes("/playlists/rate-limited/items")) {
+        return new Response(JSON.stringify({ error: { status: 429 } }), {
+          status: 429,
+          headers: { "Retry-After": "120" },
+        });
+      }
+
+      if (url.includes("/playlists/unknown/items")) {
+        throw new Error("Unknown playlist should not be probed during cooldown");
+      }
+
+      throw new Error(`Unexpected Spotify URL: ${url}`);
+    });
+    vi.stubGlobal("fetch", fetchMock);
+
+    const service = buildService();
+    const playlists = await service.getMyPlaylists();
+    const playlistsDuringCooldown = await service.getMyPlaylists();
+
+    expect(playlists.map((playlist) => playlist.id)).toEqual(["owned", "cached-accessible"]);
+    expect(playlistsDuringCooldown.map((playlist) => playlist.id)).toEqual([
+      "owned",
+      "cached-accessible",
+    ]);
+
+    const rateLimitedProbeCalls = fetchMock.mock.calls.filter(([input]) =>
+      input.toString().includes("/playlists/rate-limited/items"),
+    );
+    expect(rateLimitedProbeCalls).toHaveLength(1);
   });
 });
